+++ datanode/networkhistory/service_test.go
@@ -378,12 +378,12 @@ func TestMain(t *testing.M) {
 		log.Infof("%s", goldenSourceHistorySegment[4000].HistorySegmentID)
 		log.Infof("%s", goldenSourceHistorySegment[5000].HistorySegmentID)
 
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "QmScxLQqSrx25i5KZjeVcfEcKuRfZnYwX5HsWL18mt66Qz", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmSha1TtcRyCwXocA1rBSm5khb9wRxf4k8pF1idEKmAyAx", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmQhLCf3HCksaEHaB1Wwe6jGa5xNgnx3rDgaZq5p4RdNUS", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmQhGyknJuQzJZkde6XTR93n1wvQKvBXEHbTTFxtkrLNWH", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmWBhoG6Dp22MmZWHw8ULygciDxD2Xb9bm25LujQos7a8U", snapshots)
-		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmfVXXqPn7yhLfPLjjJecsZ17d5DSyuXuyimLpy9dsERBu", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[1000].HistorySegmentID, "Qmbhc6dntm8gZnrVVMueciLJqXH6aViaZnRCE2iTn7oi3V", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2000].HistorySegmentID, "QmXbhq694htpQP75BZTz4nDv9dreRA7Mn8aqBrbiz4yeNJ", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[2500].HistorySegmentID, "QmPExgLTRVgy7825bDC6ngWZHEe6sTgtDKZsf6owR5WhEe", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[3000].HistorySegmentID, "QmeYsvai83fNURJxU4yoEPNDvjzyx3j5gB2bXzTzbxFso1", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[4000].HistorySegmentID, "QmUqjB5chgFhPABLA4FCcT4xUXvZkGZ4wG7eMKHuZfeWhw", snapshots)
+		panicIfHistorySegmentIdsNotEqual(goldenSourceHistorySegment[5000].HistorySegmentID, "QmWvWQeEpSmM4VrihfXq6iBxwsb7HZHm7XoKUs8SdVt7C8", snapshots)
 	}, postgresRuntimePath, sqlFs)
 
 	if exitCode != 0 {