+++ CHANGELOG.md
@@ -175,6 +175,7 @@
 - [9886](https://github.com/vegaprotocol/vega/issues/9886) - Add party vesting statistics endpoint.
 - [9770](https://github.com/vegaprotocol/vega/issues/9770) - Fix `PnL` flickering bug on an already open position.
 - [9827](https://github.com/vegaprotocol/vega/issues/9827) - Expose `was_eligible` in the referral set statistics.
+- [9918](https://github.com/vegaprotocol/vega/issues/9918) - Expose vesting quantum balance.
 
 ### 🐛 Fixes